Made it possible to add negative payments.
Needed when making an invoice with a negative amount.
The usecase for this is that in denmark you are not allowed to change an invoice once
it has been filed.
So to change an invoice or cancel it, it is required to create a creditnote.
The system allows this but does not let you enter a negative payment for an invoice(creditnote in this case)
which is a mistake.