Change version number of jsoneditor from * to fixed version 5.10.0
JSONEditor will soon drop support for Bower, which has been deprecated. 

Currently, there is a dependcency on the latest version of jsoneditor `*` in the package.json file, this will break invoiceninja as soon as there is a version of jsoneditor released without support for bower. The sort term fix is to change the version number to the latest version of jsoneditor with bower support: `5.10.0`. The long term solution would be to move from bower to npm for example.

More info: 
https://github.com/bower/bower/issues/2298
https://bower.io/blog/2017/how-to-drop-bower-support/