Always Show Selected Taxes
If a person enters a custom Surcharge Label with Charge taxes selected and the label matches the regex '.+ [0-9.,]+%' e.g. 'Tax 10%' then it is always applied to the invoice even if it is 0.

Made to solve https://www.invoiceninja.com/forums/topic/quote-and-invoice-detailed-totals-view/

If you can create variable with the tax name e.g. 'Tax 10%' when taxes are created, which can be accessed from javascript I would be happy to adjust this javascript to use that variable instead of a custom label.