Run AutoBill before emailing invoice
The existing code will attempt to autobill, and follow up
by sending an invoice email that is already paid.  This seems
to be backwards.
This commit changes the order to first attempt to autobill, and
if successful will not email the invoice.  Only if configured to send
the email, and the invoice still has a balance, will the invoice be
emailed.