"Real" entity numbers in Invoice Settings preview
This refactors the GeneratesCounter trait slightly, such that arbitrary
entity numbers can be formatted according to the given padding/pattern.

With that small abstraction we can use the trait in the PdfMock instance
to show the exmple entity number using real patterns without actually
incrementing it or checking the number is available in the database.