Apply formatted numbers to design editor preview
If a custom Invoice Number Pattern is defined, this will update the mock
invoice with a formatted entity number, rather than the EAN13 from faker
in the invoice factory. This doesn't apply once a real invoice has been
created, but it can still be useful for designing during initial setup.