TemplateService - enable "raw" filter
I need to render external image in my custom invoice design. I can use `img()` function, but the output is being HTML escaped. The only fix for it is to either mark the `img()` to not escape at all, or simple enable `raw` filter. This change brings support for way more shenanigans that we can now do in custom invoice templates.

Signed-off-by: Honk <rostislavkaleta@gmail.com>